This is a refactoring of  @jrudolph's https://github.com/sbt/sbt/pull/3979 on top of test PR https://github.com/sbt/sbt/pull/4002

For every setting, we need to calculate the delegates, so `indexedDelegates` is one of hot methods.

| build                      |  sbt 1.1.1     |  1.1.2 | this patch    |
| -------------------------- | -------------- | --------------  | ------------- |
| Cinnamon (125070 settings) | 35.539 ± 0.474s  |  35.469 ± 0.318s    | 32.242 ± 0.990s  |
| Akka      (30398 settings) | 17.908 ± 0.101s  | 17.780 ± 1.494s   | 17.166 ± 0.726s  |

Given that 1.1.x branch doesn't have optimization changes merged in we can ignore the difference between sbt 1.1.1 and 1.1.x. Basically it's around 35.5s within a second of waffling.
With this patch, I am getting around 32.2, which is 3s speedup.

As you can see, with Akka's build with much fewer settings there's roughly 0.6s speedup, so this patch likely won't help you much until you have _a lot_ of subprojects.

<s>In any case, 7s would be the lion's share of Johannes's ~10s speedup.</s> (In earlier version, I had user CPU time confused with the total) I focused on just the delegation expansion, so this change is a lot more contained.