@jrudolph 2018-03-13
Thanks for taking that one for a spin and adding more tests, @eed3si9n!