@jrudolph 2018-03-13
Btw. on my 5 year old computer, Akka loads in 20 seconds, so there might be something else going on on your machine with the Akka build.