@eed3si9n 2018-03-13
> Btw. on my 5 year old computer, Akka loads in 20 seconds, so there might be something else going on on your machine with the Akka build.

Interesting. Maybe global settings causing this or global plugins (I only have `addSbtPlugin("com.jsuereth" % "sbt-pgp" % "1.1.0")` at the moment), but that's a big difference.