@eed3si9n 2018-03-19
@dwijnand Can we put this into 1.1.2, or is there a blocker?