@dwijnand 2018-03-20
I too would love to have this in 1.1.2, but given the importance of scope delegation the fact that it isn't properly tested makes me wary of merging these kinds of refactorings without even a RC.

so imo our choices are either:

* complete the retrofitting of test coverage for the scope delegation (aka #4002), or
* cross our fingers and merge this into 1.x, and then make sure we *really* test 1.2.0-RC1 to make sure we haven't broken anything

I don't, at this time, have a strong preference for or against delaying sbt 1.1.2 for this change.