@retronym 2018-03-21
> so imo our choices are either:

One approach that is useful when test coverage isn't great is to include a commit in the PR that includes the old and new implementations, and asserts the results are the same. You can run that commit through some real-world builds to check the tripwire never goes off.