@eed3si9n 2018-03-28
> Btw. on my 5 year old computer, Akka loads in 20 seconds, so there might be something else going on on your machine with the Akka build.

Finally figured this out. noob mistake on my part. Given `sbt exit  43.68s user 4.22s system 247% cpu 19.323 total` I was reading the user CPU time.