@eed3si9n 2018-06-27
Merging this into 1.x so we can test it out during 1.2 RC cycle.